requirements.txt
ccxt
ta
pandas
python-dotenv
